Move mpmath to conversion module
Reduces the size of atoms and I think organizes things better.
More smaller more-specific modules.
