+++ mathics/builtin/arithmetic.py
@@ -35,11 +35,11 @@ from mathics.core.atoms import (
     Rational,
     Real,
     String,
-    from_mpmath,
     from_python,
 )
-from mathics.core.convert.sympy import from_sympy, SympyExpression, sympy_symbol_prefix
 from mathics.core.convert.expression import to_expression
+from mathics.core.convert.mpmath import from_mpmath
+from mathics.core.convert.sympy import from_sympy, SympyExpression, sympy_symbol_prefix
 from mathics.core.expression import Expression
 from mathics.core.list import ListExpression
 from mathics.core.number import min_prec, dps, SpecialValueError
