+++ mathics/builtin/files_io/files.py
@@ -27,10 +27,10 @@ from mathics.core.atoms import (
     MachineReal,
     Real,
     String,
-    from_mpmath,
     from_python,
 )
 from mathics.core.convert.expression import to_expression, to_mathics_list
+from mathics.core.convert.mpmath import from_mpmath
 from mathics.core.expression import BoxError, Expression
 from mathics.core.symbols import Symbol, SymbolNull, SymbolTrue
 from mathics.core.systemsymbols import (
