+++ mathics/core/atoms.py
@@ -53,25 +53,6 @@ SymbolI = Symbol("I")
 SYSTEM_SYMBOLS_INPUT_OR_FULL_FORM = system_symbols("InputForm", "FullForm")
 
 
-@lru_cache(maxsize=1024)
-def from_mpmath(value, prec=None):
-    "Converts mpf or mpc to Number."
-    if isinstance(value, mpmath.mpf):
-        if prec is None:
-            return MachineReal(float(value))
-        else:
-            # HACK: use str here to prevent loss of precision
-            return PrecisionReal(sympy.Float(str(value), prec))
-    elif isinstance(value, mpmath.mpc):
-        if value.imag == 0.0:
-            return from_mpmath(value.real, prec)
-        real = from_mpmath(value.real, prec)
-        imag = from_mpmath(value.imag, prec)
-        return Complex(real, imag)
-    else:
-        raise TypeError(type(value))
-
-
 class Number(Atom, ImmutableValueMixin, NumericOperators):
     """
     Different kinds of Mathics Numbers, the main built-in subclasses
@@ -912,6 +893,7 @@ class StringFromPython(String):
         return self
 
 
+# FIXME: it would be nice to be able to move this out of Atom
 def from_python(arg):
     """Converts a Python expression into a Mathics expression.
